+++ app/controllers/tasks/people/filter_controller.rb
@@ -1,0 +1,7 @@
+class Tasks::People::FilterController < ApplicationController
+  include TaskControllerConfiguration
+
+  def index
+  end
+
+end