+++ lib/queries/collection_object/filter.rb
@@ -168,6 +168,8 @@ module Queries
       # @return [String, nil]
       attr_accessor :buffered_determinations
 
+      # TODO: See `exact[]` pattern in people
+
       # @return [Boolean, nil]
       attr_accessor :exact_buffered_determinations